Fix: Reorganised chat_service
